+++ crypto-contract-value/src/exchanges/deribit.rs
@@ -1,6 +1,6 @@
 pub use crypto_market_type::MarketType;
 
-pub(crate) fn get_contract_value(market_type: MarketType, _pair: &str) -> Option<f32> {
+pub(crate) fn get_contract_value(market_type: MarketType, _pair: &str) -> Option<f64> {
     match market_type {
         // Each inverse contract value is 1 USD, see:
         // https://docs.deribit.com/?javascript#trades-instrument_name-interval
